Financial Prediction/PersonalFinanceApp/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.0
google-auth==2.27.0

# Data processing and ML
pandas==2.2.3
numpy==2.0.0
scikit-learn==1.7.2
xgboost==2.1.3
yfinance==0.2.66
curl_cffi>=0.7.0
matplotlib==3.9.2
seaborn==0.13.2
ta==0.11.0
joblib==1.4.2
psutil==6.1.1
tqdm==4.67.1
optuna==3.6.1
beautifulsoup4==4.12.3
fake-useragent==1.5.1

# Background tasks
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Admin portal
pyotp==2.9.0
qrcode[pil]==7.4.2
sse-starlette==1.8.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0